[MSPAINT] Delete common.h and move code to precomp.h
Refactoring. CORE-19094
